# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI-Library-Backend-API",
    "description": "Backend API Document"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/ai-admin/models": {
      "get": {
        "tags": ["AI-Admin"],
        "summary": "获取模型列表",
        "responses": {
          "200": {
            "description": "模型列表"
          }
        }
      },
      "post": {
        "tags": ["AI-Admin"],
        "summary": "添加新模型",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model_name": {
                    "type": "string"
                  },
                  "model_type": {
                    "type": "string"
                  },
                  "endpoint_url": {
                    "type": "string"
                  },
                  "max_tokens": {
                    "type": "integer"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "model_config": {
                    "type": "object"
                  }
                },
                "example": {
                  "model_name": "deepseek-r1:1.5b",
                  "model_type": "ollama",
                  "endpoint_url": "http://localhost:11434",
                  "max_tokens": 4096,
                  "temperature": 0.7,
                  "model_config": {
                    "top_p": 0.9,
                    "repeat_penalty": 1.1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/api/ai-admin/models/{id}": {
      "patch": {
        "tags": ["AI-Admin"],
        "summary": "更新模型配置",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "temperature": {
                    "type": "number"
                  },
                  "is_active": {
                    "type": "boolean"
                  },
                  "max_tokens": {
                    "type": "integer"
                  }
                },
                "example": {
                  "temperature": 0.8,
                  "is_active": true,
                  "max_tokens": 8192
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "tags": ["AI-Admin"],
        "summary": "删除模型配置",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/ai-admin/models/{id}/test": {
      "post": {
        "tags": ["AI-Admin"],
        "summary": "测试模型响应",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "test_message": {
                    "type": "string"
                  }
                },
                "example": {
                  "test_message": "你好"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "测试成功"
          }
        }
      }
    },
...
    "/api/ai-admin/ollama/pull": {
      "post": {
        "tags": ["AI-Admin"],
        "summary": "拉取 Ollama 模型",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model_name": {
                    "type": "string"
                  }
                },
                "example": {
                  "model_name": "tinyllama"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "拉取开始"
          }
        }
      }
    },
    "/api/ai-admin/ollama/models/{name}": {
      "delete": {
        "tags": ["AI-Admin"],
        "summary": "删除 Ollama 服务端模型",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/ai-admin/stats": {
      "get": {
        "tags": ["AI-Admin"],
        "summary": "获取 AI 使用统计",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统计数据"
          }
        }
      }
    },
...
  }
}